refactor: split coach summary ownership and settings modules
