+++ frontend/src/features/coach/components/CoachPageLayout.tsx
@@ -42,14 +42,11 @@ export function CoachPageLayout({ apiBaseUrl }: CoachPageLayoutProps) {
     () => isAvailabilityConfigured(settingsContext.settings?.availability),
     [settingsContext.settings?.availability],
   );
-  const cachedSummary = selectedItem ? sessionCache.getSummary(selectedItem.id) : null;
 
   const chat = useCoachChat({
     apiBaseUrl,
     workoutId: selectedItem?.id ?? null,
     aliasRange: workoutList.visibleWeekRange,
-    cachedSummary,
-    onSummaryLoaded: sessionCache.setSummary,
   });
   const hasSettingsLoadError = Boolean(settingsContext.error);
   const chatError = isAvailabilityRequiredChatError(chat.error) ? null : chat.error;
@@ -76,13 +73,6 @@ export function CoachPageLayout({ apiBaseUrl }: CoachPageLayoutProps) {
     setShowConfirmWithoutChat(false);
     setIsEditing(!(selectedItem?.summary?.savedAtEpochSeconds ?? null));
   }, [selectedItem?.summary?.savedAtEpochSeconds, selectedWorkoutId]);
-
-  useEffect(() => {
-    if (chat.summary) {
-      sessionCache.setSummary(chat.summary);
-      workoutList.replaceSummary(chat.summary);
-    }
-  }, [chat.summary, sessionCache.setSummary, workoutList.replaceSummary]);
 
   const isCurrentSelection = useCallback((workoutId: string) => {
     return selectedWorkoutIdRef.current === workoutId;
