+++ frontend/src/features/coach/hooks/testRender.tsx
@@ -1,0 +1,20 @@
+import { renderHook, type RenderHookOptions, type RenderHookResult } from '@testing-library/react';
+
+import { CoachSessionCacheProvider } from '../context/CoachSessionCache';
+
+type HookCallback<TProps, TResult> = (props: TProps) => TResult;
+
+export function renderCoachHook<TResult>(callback: () => TResult): RenderHookResult<TResult, undefined>;
+export function renderCoachHook<TProps, TResult>(
+  callback: HookCallback<TProps, TResult>,
+  options: RenderHookOptions<TProps>,
+): RenderHookResult<TResult, TProps>;
+export function renderCoachHook<TProps, TResult>(
+  callback: HookCallback<TProps, TResult> | (() => TResult),
+  options?: RenderHookOptions<TProps>,
+): RenderHookResult<TResult, TProps | undefined> {
+  return renderHook(callback as HookCallback<TProps | undefined, TResult>, {
+    wrapper: ({ children }) => <CoachSessionCacheProvider>{children}</CoachSessionCacheProvider>,
+    ...options,
+  });
+}
