+++ src/adapters/mongo/settings/tests/mod.rs
@@ -1,0 +1,4 @@
+mod bootstrap;
+mod mapping;
+mod repository;
+mod support;
